fix: use an atomic chunk+offset cursor
Previously the chunk index wasn't atomic, so there was a very small
chance (where there was sufficient space in the previous chunk to allow
a small allocation through after a chunk resize was triggered by a
slightly larger allocation), that concurrent updates could result in a
data race. The arena now uses a 64-bit atomic with bit shifting to model
the chunk+offset, avoiding this.